First upgradable GuniLev via proxy
This version is not tested rigorously yet, but I assume that the overall architecture is sound. It preserves the new LP creation/deletion properties of the earlier contract as well as keeping store of user ilks. This aforementioned logic is taken care of in the proxy contract, whereas specific wind/unwind logic is left to be upgradable in the implementation contract. Stylistically, I have arranged all shared variables and functions before the contructor/initializer. I assume that Proxiable provides sufficient storage clash protection, but it is important that all future implementation contracts inherit from it.